Fix typos & update information user is given
I've removed the date of testing and the various setups which were
tested. This was something I never foresaw being such a pain and very
irritating and time-consuming to maintain, especially with a bazillion
other projects. I simply don't have the time nor energy to do that in
my own stead.

However, with simplify-ubuntu's updated information the user sees, such
as the warnings, plus all of the portability changes, I believe it is
now far better than it ever was.

Provided the user is aware that he or she _must_ watch out for packages
which their distribution or configuration will _also_ want to remove;
this is very important!

For someone who knows what they're doing, it's nothing major to have to
reinstall a few packages accidentally removed alongside something the
user didn't want, but a newbie could get bitten by this quite badly if
no due attention and care were given throughout.