Reinforce MAIN(), clean deps, & add curl support
Since `curl` is so often a choice-downloader for Linux developers and
probably non-regular users alike, I've added support for it.

The `wget` functionality is still available, should `curl` not be
found. They both use similar options, so _should_ yield the same.

Removed what I believe is a no-longer-relevant `wget` check.

Correctly (?) `$COL` to `$COLUMNS`, although I doubt it'll matter, as
`tput` seems to be the only viable option when within a script, at
least, with your typical installation of Linux.

Removed `tput` as a main dependency, but is still checked after the
main dependency `for` loop, _if_ `$COLUMNS` is null.