Greatly improve base functionality
Less skeletal at this point. Both page-by-page and the kind you scroll
are now available.

The main thing I'm wanting now is for the line pushing to work
correctly (keeping cursor at the bottom of the screen) but it's a
little tricky, as this code is quite complex.

I would say it's at this point that you could use `shlides`, but expect
bugs and consider it as more of a beta, or even alpha.