Replace `PushLines()` with `CursorDown()` (tput)
Instead of the lengthy and complicated `PushLines()`, I resolved to
just using `tput`. I must have gotten so stubborn trying to do it with
`bash` itself, perhaps looking for that challenge. Lol

This way is probably a lot quicker and more reliable anyway, not to
mention makes the code far less complex, and a bit shorter. I will do
the same with the title eventually, but it messes with other stuff, -
due to the way in which I've programmed the borders.

Now I can focus on more interesting things than cursor management!