Improve footer and catch too few footer columns
Now when the footer columns exceed a certain amount, relative to your
current terminal's number of columns, you'll be given in an error
telling you as much, and then `shlides` will quit. This is to avoid
buggy output and the tackiness it brings.

Removed some now-unused code which was relevant to past method(s).

Added missing semi-colon to `tput` color assignments.