Overhaul the text alignment (non-body)
Due to bug after bug relating to alignment header and footer, I decided
it would be well worth just overhauling how it all worked.

Now it doesn't matter if the columns are odd or even, and it shouldn't
matter how short or long the values are. It's pretty much all done via
`tput`, which saves a lot of hassle, and confusing `printf` instances.