Improve border handling (especially non-ASCII)
I've made a note of 3 fonts I found to be very good with `shlides`.

  FreeMono
  Monospace
  Oxygen Mono

Effectively, if it's a good monospace font, it'll probably work
well, but this applies to all text on the terminal, not just here.

Instead of spreading out ASCII and non-ASCII character assignments
throughout `shlides`, all these characters are set early on, which also
makes any future changes much easier and the code cleaner.

There are now two types of side-bars (previously `SB`), in the form of
`SBL` (left-leaning) and `SBR` (right-leaning). This allows for more
variation, and support for non-ASCII characters with special alignments
for each side.

The new border style (characters used, not layout) is now a better fit
for more fonts, and is used in many other terminal programs.