Add `cito`
This is basically insit's successor, without the fluff. It's got decent
logging, it's lightweight and portable, it makes use of the new library
file (`libtflbp-sh`) and is a _lot_ easier to maintain.

There's a basic file removal (akin to _uninstalling_) feature right
now, but I may expand on it at a later date.

Don't worry about installing this manually. I'll have instructions in
the next commit, regarding this repository's `README.md` file.