Add interpreter auto-detection to AutoExec
Also added `--once|-o` to bypass the execution loop, which is there
mostly for debugging reasons, but there may be some weird case for it.

The man page was also updated to reflect the new changes, and the Bash
completions script was tweaked to allow for regular file completion, so
the user can complete his or her way to an appropriate file.