Remove `versions` stuff; was barely even used
I'll probably re-incorporate something like this again, in the future.
For now, however, I think it's just extra hassle with which to deal, -
and an extra opportunity for silly & forgotten commits.

The few shell program currently still using the `versions` file, won't
work with the `--update` option(s), but it shouldn't affect the rest.