Add more to `tfl-ubuntu-setup` & fix Cito
Cito has had this irritating bug for who knows how long! I've fixed it
now, finally. It was causing files to incorrectly be accepted as though
they had a Cito line, causing all manner of silliness behind the
scenes. It should be fine now.

As for `tfl-ubuntu-setup`, it now installed VimConfig and i3Config.