Lower restrictions of log file & tweak args
Updated the man page as well, with better, more standardized formatting
and some newer information, such as to NOTES.

I've cleaned up the argument processing ever-so-slightly, primarily the
user's experience thereof, in that they will only see their option
provided (during `Err()` call) rather than any and all within that
group. Example:

```
$ cito -H
ERROR: Option '-H' requires an argument -- see: cito -h
```

Root access is no longer required when viewing the log file, finally. I
probably should've done this ages ago, but it didn't occur to me that
the permissions for the log file were a bit excessively restrictive.

On the flip-side, and likely from here-on-out, the act of a user
viewing the log file is no longer itself logged.

Cito shall endeavor to maintain a mode of `644` for the log file,
instead of the previous `600`, allowing all to read the log file. If
this is a concern, for you, please let me know, so I know to release an
optional patch.

When viewing the log file with Cito, temporary files will no longer be
created, therefore nothing needs to be removed via the signal trapping
function.

Hashes have been updated.