Fix up mif's help output a bit
Eventually, this and similar programs will take the more traditional
approach, while most of this information would be reserved for a man
page.

To some extent, however, MIF is a personal program. I actually don't
use it anymore, as I don't tend to watch movies on the PC, but I would
still recommend this to those who're looking to watch films on their
computer, but keep their collection organized and indexed.

The main thing is to know what you've actually seen, so that when you
go to watch a new film, you don't have to perform memory gymnastics to
recall ones you've already watched. It's a quality-of-life program.