Remove Cito line from tfs (due to simplification)
Not comfortable having this installed in '/usr/bin' with it being so
simple now. It would work no differently, but I have an arbitrary set
of standards for a program to be worthy of '/usr/bin', and this no
longer meets that; it'd be good in '~/bin', however.

Also, in the last commit, I added gnome-screenshot support, which I
forgot to mention -- oops.