Add SignalHandler() & `--keep-files|-k` OPTs
GLKFU will handle premature exits, now. Unfortunately, this won't work
with everything, but it'll definitely be a better experience.

This can be overridden with the new flags though, in-case you have
issues or just wish to keep the files, such as if you wanted to
download the latest Linux tarball but not actually do anything with it.