Add `--list-kerns|-l` to glkfu for glkfu-list use
Since GLKFU itself is using BASH, I'd certainly prefer that
`glkfu-list` also use BASH, but the data parsing and version sorting is
just too nice a feature to pass up.

I think I could figure something out with sort(1) for a BASH version of
`glkfu-list`, but at least the functionality is there in place for now.