Remove catmedia; no longer needed
Decided to remove this because I don't think it's worth using. It's no
problem at all to use FFMPEG directly to concatenate video files. Using
an alias or a simple function for this makes life even easier.

```bash
clips ()
{
    if ! [ $# -eq 0 -o $# -eq 1 ]; then
        printf "file '%s'\n" "$@" > flist;
        ffmpeg -v 0 -stats -f concat -safe 0 -i flist output.mp4;
        rm -v flist;
    else
        printf 'ERROR: No video clips provided.\n' 1>&2;
        return 1;
    fi
}
```

The above, for example, is what I use for the Learn Linux channel.