Add bug TODO to Cito & tweak an Err() call
Using:

    ERROR: Root permissions required in order to log activity.

Instead of:

    ERROR: Root permissions required for system-wide changes.

This is because there is no error, per the latter version, for any
operation which does require root privileges, even if logging is
disabled.

So, instead, I've opted for the user simply being told a file failed to
install; it leaves a little to guess-work, but it saves confusion the
old error message might've given when the user disables logging.

The log can always be used for more information, anyway.