Update & optimize to no longer require LSPKG
Plus tweaked welcome message, Usage(), and some code in general. The
best part is by far the new code allowing all packages to be scanned
without the use of LSPKG, and it's much, much more efficient, in that
it only scans the package database ONCE, rather than every single time
for each search, which was the old behavior.

Simplify Ubuntu should now be MUCH faster.

I've also added output so the user can finally see for which packages
are being searched and which ones are actually found; should really
help the user understand what's going on.

I've gathered the functions together to help with readability.
Generally just tidied up the code a bit; not huge changes, but it made
a notable difference.