Check extension instead of file(1) for C/C++
It seems that file(1) was just checking the file extension anyway, so
it didn't make any sense to me to use file(1) to do what BASH can
itself do.

AutoExec will check for '*.c' and '*.cpp' (C++) files. It might be nice
to have some sort of fallback procedure, for when gcc(1) fails to
compile it, determining the file is not even a C file, because some
might weirdly name their scripts or whatever files '*.c' or '*.cpp'.

I might simply make it so the user can disable the C/C++ functionality
altogether, for just such a situation, by using a flag.