Remove all signal trapping in GLKFU
Updated manual page to reflect these changes.

The reason for removing this flag and functionality is because it was
ultimately annoying and unnecessary.

There are times when the user will probably want to exit, such as if
they wish to use GLKFU to acquire the desired kernel, maybe have it do
some stuff, _then_ do their _own_ thing.

If, however, the user forgets the `-k`|`--keep-files` flag, as I've so
often done, he or she winds up losing all the files GLKFU downloaded.
If a user has a particularly slow or otherwise limiting Internet, it'd
be immensely frustrating.

Lastly, it introduced new problems, simply put.

Note, that the files will _still_ be removed, or prompted for as
much, but no signals catching will result in the files going _poof_.
I, for one, am hella pleased with this commit!