Add compound-demo.sh concept script
Not really a program or utility, just something I've put together after
a _lot_ of thinking and trying to address the lack of compounded
short-form argument processing in my programs, both Shell and Perl.

Suggestions to address problems mentioned in the comments of this file
(not PRs) are more than welcome. TIA.