Remove sort(1) and support movestack & setlayout
Those functions now support the use of arguments.

I decided to remove the sorting, because I don't think it's all that
useful, and it hadn't occurred to me that the user may well have orderd
their bindings in a specific way. It might come back as an optional
feature, but I'm on the fence right now.

Regarding the new bug report TODO item: #7