Offer to use existing tarball & confirm it
What I mean by "confirm it" is that GLKFU will now check that the
downloaded kernel is the one the user desired. This is largely done
with `make kernelversion`.

Offering to use an existing kernel is something I've personally wanted
for ages, so I'm happy to finally include that functionality.