Work around unreliable configuration file code
The old code made sense at the time, but the longer I used GLKFU, the
more apparent it became that such a method wasn't reliable. Instead, as
a temporary measure, I've opted for the user to be prompted to either
use the configuration file for the current kernel, or to get a valid
file path via STDIN. A feature TODO has been made regarding this, as
it's certainly not ideal.