Minor optimization of capitalization
Avoiding the need for BASH arrays and the `for` loop, during the
capitalization of the first character of each word in the binding. I
did this to reduce the complexity of the code, and hopefully to
ever-so-slightly improve the performance.

My brief testing shows this works just fine, but if there are any
issues, be sure to let me know!