Simplify interpreter usage (does it better)
I found out that AE wasn't using the interpreter properly, and realised
as a result that the way in which the interpreter was used wasn't even
correct. AE should use the full shebang, minus the leading `#!`, which
is what this commit aims to correct, hopefully without breaking stuff!

I've removed the Interpreter() function, making it easier to see when
the `Exec` variable is assigned a value.