Move TryBASH to 'devutils' & add test-bash.sh
TryBASH is no longer a 'program', but a 'devutils'-worthy script.
Basically, it's so niche and for my personal testing uses, that I
figured it made more sense to put it here. It's also no renamed to
'compile-bash.sh'.

This script will no longer throw you into the selected version of
BASH, as it's now just a quick and EZ way to compile it.

The 'test-bash.sh' script is for the actual testing. I might combine
the two eventually, but this'll do for now.