Compile all (non-alpha/beta) versions of BASH
Finally, this script does what I now want. Drawback? Determining the
order of versions is now going to be a huge pain, because of the
variety of version strings. I'm thinking of parsing the site's
timestamps for each tarball, but I doubt that'll be too reliable, as
a tarball could've been changed post-release.