Add simplified.patch for AE & patch README.md
First AE patch! Hope it helps someone. Remember, this only changes AE
itself, _not_ the man page.

Because of how rarely I actually need to change AE, these days, I felt
it appropriate to introduce patching, for some different takes on AE, -
without impacting the original code.

It's possible that, from now on, only patches will be added for AE.
It's also possible that patches will break as of a given new version of
AE, so I'm afraid there are no guarantees.

The patch included in this commit is for at least AE 2022-01-11.