Add new LUDL project (Linux Unplugged Downloader)
Useful wee tool. It's very new, of course, so bear with me. Seems to
work well, but please let me know if you find any bugs. I'd like this
to eventually be of a similar quality to WCDL (in PerlProjects).

If you're part of Linux Unplugged and have any suggestions, issues, or
concerns, please don't hesitate to let me know.

An installation script and man page will probably be added soon.