Add bug TODO about associative array randomness
Sorry. This was just an oversight, plain and simple. I forgot
associative arrays are randomly ordered in BASH.

I'll fix this in the same way I did with CSi3, by prefixing a number to
each podcast key, then sort them sequentially. Either that or I'll
somehow manually sort them with bubble sort, but the former is an
easier and faster solution, I believe.

A feature to select a random podcast would be awesome, though.