Remove AE patch and never diff(1) patch(1) again
I'm so done with the idea of patching, at least with patch(1); the
slightest change seems to mess up a patch, meaning I have to re-create
it. Why do people do this to themselves?!

I really liked that patch. Maybe I'll create a separate branch for
it, but then I still have to maintain both versions. BAH!