Improve and simplify filesitter
It'll now only work with 1 file, to keep things simple. It'll also no
shut down or suspend the system, because the user themselves can
execute that command very easily, such as:

```
filesitter -i 10 ~/.bashrc; systemctl suspend
```

Instead of checking for the size of the file, seconds since Epoch is
used, in order to detect whether the file has been written to, as
opposed to just an actual size change.