Correct build path error showing unnecessarily
This issue was largely down to GLKFU's build path check being placed
far too early in the code. I had to move it much further down to allow
the user to list kernels and 'just check', or they'd constantly be
told the build path is invalid, if it is, despite it not then being
pertinent.

I've updated the comment header accordingly.