Mostly rewrite and overhaul freelfm
A million times better. Night and day difference. It took ages, because
it was so bloody far from my current standards. Just need to do a few
more things before I push and merge dev with master.

I don't use that music site and the music isn't really up my street, so
this is something I initially wrote and currently am working on for
whomever else _does_ care for it.

This rewrite also traps, unlike the original, so it cleans up after
itself and correctly uses a temporary directory in '/tmp' for web page
downloading.

I realise I needn't sort with bubble, of all things, but the first
page doesn't list every single page, so performance shouldn't be a
concern. I prefer to keep things in the language I'm actually using.