Remove LTS pointers (unreliable) & tweak sorting
Further simplified `Compare()` to avoid doing that which in this case
is unnecessary. If I use `split()`, I don't need to test for periods, -
for example. Also changed the comment above the function definition
for clarity and to reflect the changes made to the function in this
commit.

The LTS pointers seemed to not even be working. Rather than fixing
this, I decided to remove the code for that feature, at least for the
time being. In all likelihood, if the user is compiling their own
kernels, they probably already know or are willing to find out which
kernel is an LTS.