Handle small terminal & reorder trapped errors
The ordering is now sane, where before the order didn't make much
sense. For example, the check for the file error was performed _after_
the check for a valid shebang or found interpreter.

A hard minimum of 30 columns and 2 lines is now in place to avoid the
mess of output that would previously ensue if this minimum columns were
breached, and to avoid any unexpected behavior with too few lines.