Resurrect and improve ROKs (finally!)
I'm using this on my system from now on, as it's such a timer-saver.
I just don't know how it'd fair with distribution flavor kernels and on
different distributions. It really comes down to linux-version(1), and
its availability, I think. If you're willing to test this, please do!
Let me know how it goes via the Discussions tab on GitHub.