Fix bug preventing re-detection of shebang (duh!)
I have absolutely no idea how this _glaringly_ obvious solution didn't
scream at me years ago! It was checking to see if `$Exec` was empty, -
before processing the shebang. Since the `$Exec` isn't empty after that
first detection, it never got a chance to re-process the shebang. A bit
embarrassing, but at least it's fixed now. This also is correctly
overridden by the `-e` and `--exec` flags, so this commit shouldn't
break that. I'm so happy this is _finally_ fixed.