Add `-n`|`--number` && tweak final loop
The final loop needed tweaking to add the line number functionality, -
but this was also made to be clearer, so it's easier to understand
what's going on in this admittedly confusing section.

The new flags allow the user to display the line number on which each
key binding is discovered in the configuration file, which may become
useful for reference. By default, this is displayed in green and is
dynamically displayed in its own column.