Update CSi3 man page with `-n` info and NOTES
The new NOTES section was much needed. Hopefully now things will be a
bit clearer for users. Since the information found via `--help` is now
in and expanded upon in the man page, I've opted to remove it from the
usage output.

I've also improved the description for CSi3, and updated any mention of
gaps, following the merge (finally!) of Airblader's `i3-gaps` into the
official i3WM.