Allow CSi3 to parse ALL variables (see #10)
Updated the 'README.md' and man page to reflect these changes. There's
a new flag which disables this functionality, as explained in the man
page and the comments in the new code included in this commit.

Admittedly, I'm not _yet_ 100% sure this is stable, but I've no reason
to expect it to fail, it's just that there are a lot of moving parts.
Testing showed this works fine. I'm actually amazed not only how well
it works, but how fast BASH can manage to do all of this.