Add support for split up lines to CSi3
Very happy with this feature. It was a pain to incorporate. It works
with multiple lines, not just a line split up once. Regarding the line
numbering, the original line (start point) is the one counted as the
line on which the binding was found, which I feel makes the most sense.