Properly handle `--release` bindings (Issue #17)
The NOTES section of the man page has been updated, and I've corrected
the dependency in the header. I recently incorporated name
referencing, but forgot that would require a newer version of BASH.

I'm not overly keen on the approach to these `--release` bindings, but
it's tidy and it works. I could have left this issue sat there for all
eternity, having never incorporated it into CSi3, but in the end, I
think including it in some way is better than not adding it at all.

Ideally, I'd have _some_ sort of special character (I like the caret)
on the _left_ side, where the bindings are, but I simply couldn't
justify doing that, because it would look strange and be confusing to
the user. It just wouldn't look right. One solution would be to
incorporate color, but then what if the user is colorblind, has colors
disabled, or CSi3 is operating without interactivity (e.g., a pipe)?
That wound up being a dead-end, as well.

Next, I'd like to add a flag which allows the user to filter by these
release bindings.