Optimize with new parser to get latest versions
To clarify, this better gets the version _strings_.

This uses RSS feeds, which are quicker to load and quicker to parse.
The downside is the added complexity, but it's not too bad. The bonus
is now there are two methods, adding extra reliability. If there is an
error fetching the RSS data, the fallback method will be immediately
used. Lastly, this is probably kinder to the server(s).