Improve and rename filesitter to fsit
Now simply focuses on doing nothing until the file is touched. This
should be very convenient when doing something on the command-line, or
in scripts, but not wanting to proceed until the file has been written
to.

This is a good candidate for a rewrite in POSIX shell, or perhaps
Python. Will consider it, but BASH is doing this very well, and this is
portable.