Use one-line progress & process existing file
Using hashing for checking the file, so it should be fast enough.

Essentially, this allows you to reuse an existing file, rather than
starting from scratch every time. This will allow you to do all this in
different sessions, and allow you to essentially maintain the file.

The progress is all together in a single line, which looks very nice
and doesn't fill your terminal up with noise. Much more presentable.
I've also added a check for interactivity, so this output isn't
displayed if inactive.