Avoid wrapping if listing dependencies to install
Also tidied the code a little, by removing a nest caused by `case`.

I've avoided wrapping when listing packages, in-case the user has
issues copy-pasting and winds up doing something wrong by mistake. It's
unnecessary, especially given the target audience of this tool in the
first place, but I figured it was better safe than sorry.

I've added dim grey to the `$` or `#` when displaying a command the
user can run, to denote that it's different, so as to hopefully suggest
that you don't _actually_ run that part. Not everyone is familiar with
that syntax.