Ditch slow complex code & fix missing chars bug
As much as I loved what was there, as nice as the 3.1 compatibility
was, and as hard as I worked to discover that awesome approach, it was
absurdly difficult to maintain and unpleasant to work with. A bug
causing certain characters to not be shown was just frustrating to try
to solve.

The downside to this commit is that Snotes now requires 4.1 instead of
just 3.1. A shame, but given <= 4.3 is vulnerable to ShellShock, nobody
in their right mind should be on such an old version of BASH in the
first place. The plus side to this commit, is speed and reliability!

As it happens, I've just realised I broke that 3.1 portability at some
point, with the addition of `+(...)` in `[[`. I guess it's just as well
I switch to a 4.1 dependency.