Improve handling of failed version comparisons
This improves the initial use of GLKFU, where a user would typically
have a _generic_ kernel, which wouldn't compare properly with a
standard upstream (_stable_) kernel, like 6.4.3.

Prior to this update, a weird situation like this might occur:

	```
	$ glkfu
	* Finding latest kernel ... [OK]
	* Verifying tarball filename ... [OK]
	Latest Linux version: 6.4.3
	Current Linux version: 6.1.0-9-amd64
	Unable to determine if latest kernel is newer.
	Q: Continue with GLKFU? (Y/N) y
	Kernel appears up-to-date -- quitting.
	```

That was on Debian 12. With this commit, it should just continue, but
still prompt the user.