Remove trimit
Looking at the code, I realise it could actually be problematic. Unless
I'm having a blonde moment, it doesn't compensate for './file', it
would be better if it had an option for operating on a directory, and
it doesn't account for 'file'. Not worth keeping around, in-case it
causes problems for anybody.

I also believe it would be much cleaner to use `while` loops.

Example:

```bash
for Orig in "$@"; {
	while [[ ${Orig:0:1} == [[:space:]] ]]; do
		File=${Orig:1}
	done

	while [[ ${File:${#File} - 1:1} == [[:space:]] ]]; do
		File=${File::-1}
	done

	mv --no-clobber "$Orig" "$File"
}
```

That's not a robust approach, as it currently is, but it's a start.